hawk-scanner-v2/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
arq>=0.26.0
redis>=5.0
dnspython>=2.6.0
anthropic>=0.40.0
